+++ src/Interpreters/SelectQueryOptions.h
@@ -70,8 +70,6 @@ struct SelectQueryOptions
       */
     bool merge_tree_enable_remove_parts_from_snapshot_optimization = true;
 
-    bool skip_optimize_redundant_functions_in_order_by = false;
-
     SelectQueryOptions( /// NOLINT(google-explicit-constructor)
         QueryProcessingStage::Enum stage = QueryProcessingStage::Complete,
         size_t subquery_depth_ = 0,
